test: PerformanceService 단위 테스트 코드에서 mock 데이터 생성 코드 외부로 분리
